Refactor integration test
