+++ tests/integration/test_e2e.py
@@ -1,14 +1,13 @@
 #  Copyright 2024 Canonical Ltd.
 #  See LICENSE file for licensing details.
-import pytest
+from typing import AsyncIterator
 from github.Branch import Branch
 from github.Repository import Repository
 from juju.application import Application
 from juju.model import Model
-from openstack.compute.v2.server import Server
-from openstack.connection import Connection as OpenstackConnection
+import pytest_asyncio
 
-from charm_state import TOKEN_CONFIG_NAME
+from charm_state import InstanceType
 from tests.integration.helpers.common import (
     ACTIVE,
     DISPATCH_E2E_TEST_RUN_WORKFLOW_FILENAME,
@@ -16,31 +15,37 @@ from tests.integration.helpers.common import (
     dispatch_workflow,
     reconcile,
 )
-from tests.integration.helpers.openstack import setup_repo_policy
+from tests.integration.helpers import openstack, lxd
 
-
-async def test_end_to_end(
+@pytest_asyncio.fixture(scope="function", name="app")
+async def app_fixture(
     model: Model,
     basic_app: Application,
+    instance_type: InstanceType,
+) -> AsyncIterator[Application]:
+    """Setup and teardown the charm after each test.
+
+    Ensure the charm has one runner before starting a test.
+    """
+    if instance_type == InstanceType.LOCAL_LXD:
+        await lxd.ensure_charm_has_runner(basic_app, model)
+    if instance_type == InstanceType.OPENSTACK:
+        await openstack.ensure_charm_has_runner(basic_app, model)
+    yield basic_app
+
+async def test_e2e_workflow(
+    model: Model,
+    app: Application,
     github_repository: Repository,
     test_github_branch: Branch,
 ):
     """
     arrange: An app connected to an OpenStack cloud with no runners.
-    act:
-        1. Change number of runners to one and reconcile and run check-runners action.
-        2. Change number of runners to zero and run check-runners action.
+    act: Run e2e test workflow.
     assert:
-        1. One runner is spawned.
-        2. No runners exist and no servers exist on openstack.
     """
-    # 1.
-    # Waits until one runner is spawned.
-    await basic_app.set_config({"virtual-machines": "1"})
-    await reconcile(app=basic_app, model=model)
-
     await dispatch_workflow(
-        app=basic_app,
+        app=app,
         branch=test_github_branch,
         github_repository=github_repository,
         conclusion="success",
